--- a/app/src/lib/git/index.ts
+++ b/app/src/lib/git/index.ts
@@ -1,4 +1,3 @@
-export * from './add'
 export * from './apply'
 export * from './branch'
 export * from './checkout'