🌵 **This depends on #1403** 🌵 

**TL;DR: This cuts the time it takes to make a commit with ~900 unstaged files from 1 minute to 4 seconds**

As I was working on #1403 I tested committing using a repository with ~900 unstaged files. I created the repository by throwing in some node modules

```shell
git init lotsa-files
cd lotsa-files
npm init
npm install --save cheerio request got
```

That gave me 917 unstaged files to work with. Committing all of these in one swoop in Desktop took about 1 minute and 2 seconds on average while `git add .` took ~4s.

The reason for it taking so long was that we call out to `git add` for each added/modified/removed file. On Windows this incurs a _significant_ cost so I switched over to using `update-index` instead which lets you pass a list of files over stdin, allowing us to update the index in bulk. This cut the time down to 4.6 seconds on average, shaving almost a full minute from our commit time.

Now, a change like this isn't uncontroversial and I believe it should fall under the same question that was posed in https://github.com/desktop/desktop/pull/1397#issuecomment-299890017. One thing that speaks in favor of this PR landing (aside from the performance boost) is that the behavior we currently have implemented using add is unintuitive at best.

Please bear with me as I run through a fairly contrived example of a renamed file.

```shell
echo "foo" > foo && git add foo && git commit -m 'foo'
git mv foo bar
echo "bar" > foo
```

After this both apps show the following (I've only selected the rename change)

![image](https://cloud.githubusercontent.com/assets/634063/25817431/96b86dea-3427-11e7-9c0b-739e23b1418b.png)

But after committing this change the results differ

![image](https://cloud.githubusercontent.com/assets/634063/25817585/18a54cc4-3428-11e7-8157-42eb7c44dd2f.png)

The old version (using `git add`) commits the change file even though it wasn't selected, whereas the new version (using `git update-index`) correctly commits only the rename and leaves the modified source file untouched in the working directory.

![image](https://cloud.githubusercontent.com/assets/634063/25817642/400cf44c-3428-11e7-8f55-ff00403166d1.png)